Use fluent ui control for tabs
