+++ packages/playground/src/react/output-tabs/output-tabs.module.css
@@ -1,0 +1,18 @@
+.tabs {
+  border-bottom: 1px solid var(--colorNeutralStroke1);
+}
+
+.tab-divider {
+  flex: 1;
+  border-right: 1px solid var(--colorNeutralStroke1);
+}
+
+.tab {
+  height: 40px;
+  border-right: 1px solid var(--colorNeutralStroke1) !important;
+  border-radius: 0 !important;
+}
+
+.tab--selected {
+  background-color: var(--colorNeutralBackground5) !important;
+}
